Cleanup, make ConfigParser an interface
