+++ src/main/java/net/minestom/server/config/BuilderImpl.java
@@ -1,16 +1,18 @@
 package net.minestom.server.config;
 
-class BuilderImpl implements Config.Builder {
+import org.jetbrains.annotations.NotNull;
+
+final class BuilderImpl implements Config.Builder {
     private int compressionThreshold = 256;
 
     @Override
-    public BuilderImpl compressionThreshold(int compressionThreshold) {
+    public @NotNull BuilderImpl compressionThreshold(int compressionThreshold) {
         this.compressionThreshold = compressionThreshold;
         return this;
     }
 
     @Override
-    public Config build() {
+    public @NotNull Config build() {
         return new ConfigV0(0, compressionThreshold);
     }
 }
